+++ src/model-id.js
@@ -7,5 +7,5 @@
   else (root.ESS = root.ESS || {}).MODEL_ID = factory();
 })(typeof globalThis !== 'undefined' ? globalThis : this, function () {
   'use strict';
-  return '8743689443f26cffa5c15bbbc1c12fba18d6276f9c00fc56f1a5831699edda9f';
+  return 'cd6670acbfd824993df261f4bfda726ebaba893103cb69dbb8249f09fa327332';
 });